Event page refactoring

+++ services/experimentation-service/internal/model/graph/graphTimeResult.go
@@ -4,5 +4,5 @@ import "time"
 
 type TimeScaleDataPoint struct {
 	Time  time.Time `json:"time"`
-	Value int64     `json:"value"`
+	Value int       `json:"value"`
 }
